# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "McDee Platform API",
        "description": "Multifunctional Marketplace & Services API",
        "contact": {
            "email": "support@mcdee.test"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Local Server"
        }
    ],
    "paths": {
        "/api/apartment/bookings": {
            "get": {
                "tags": [
                    "Apartment Bookings"
                ],
                "summary": "Get my apartment bookings",
                "operationId": "1024bba4fdcc966175c70c8f1e093e46",
                "responses": {
                    "200": {
                        "description": "List of bookings"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Apartment Bookings"
                ],
                "summary": "Book an apartment",
                "operationId": "b9e7c7b2de839b4bcab677e31f7bba87",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "apartment_id",
                                    "check_in",
                                    "check_out"
                                ],
                                "properties": {
                                    "apartment_id": {
                                        "type": "integer"
                                    },
                                    "check_in": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "check_out": {
                                        "type": "string",
                                        "format": "date"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Booking created"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
        "/api/ride/location/update": {
            "post": {
                "tags": [
                    "Ride Hailing"
                ],
                "summary": "Update GPS location for a ride",
                "operationId": "587cc0564bda4928df72e00080d33dfc",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "ride_id": {
                                        "type": "integer"
                                    },
                                    "lat": {
                                        "type": "number"
                                    },
                                    "lng": {
                                        "type": "number"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "GPS logged"
                    }
                }
            }
        },
...
        "/api/ride/request": {
            "post": {
                "tags": [
                    "Ride Hailing"
                ],
                "summary": "Request a ride",
                "operationId": "28fe8f8ee86fabf1561b68865f92b14b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "pickup_lat": {
                                        "type": "number"
                                    },
                                    "pickup_lng": {
                                        "type": "number"
                                    },
                                    "dropoff_lat": {
                                        "type": "number"
                                    },
                                    "dropoff_lng": {
                                        "type": "number"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Ride created"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/ride/update-status": {
            "put": {
                "tags": [
                    "Ride Hailing"
                ],
                "summary": "Update ride status",
                "operationId": "03ac2eea29ca76ab96272c27231e3320",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "ride_id": {
                                        "type": "integer"
                                    },
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "on_trip",
                                            "completed",
                                            "cancelled"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Status updated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/ride/history": {
            "get": {
                "tags": [
                    "Ride Hailing"
                ],
                "summary": "Get ride history",
                "operationId": "71dc7f2b11e22a6e1f6e601263023896",
                "responses": {
                    "200": {
                        "description": "List of rides"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/ride/online": {
            "post": {
                "tags": [
                    "Ride Hailing"
                ],
                "summary": "Set rider online",
                "operationId": "62fabb36cf75595ba17def019828a738",
                "responses": {
                    "200": {
                        "description": "Rider is now online"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/ride/offline": {
            "post": {
                "tags": [
                    "Ride Hailing"
                ],
                "summary": "Set rider offline",
                "operationId": "6a063441e2de6453535bf7870eaf2d12",
                "responses": {
                    "200": {
                        "description": "Rider is now offline"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/ride/rate": {
            "post": {
                "tags": [
                    "Ride Hailing"
                ],
                "summary": "Rate a completed ride",
                "operationId": "44ab5cf6e8e89bec8e07419341098e7f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "ride_id": {
                                        "type": "integer"
                                    },
                                    "rating": {
                                        "type": "integer"
                                    },
                                    "comment": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Rating submitted"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
...
}